tests/wpt/tests/css/css-animations/parsing/animation-trigger-type-valid.tentative.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Animations: parsing animation-trigger-type with valid values</title>
<!-- TODO(crbug.com/390314945): Replace with spec link when the spec lands. -->
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/8942 ">
<meta name="assert" content="animation-composition supports the full grammar '<single-animation-trigger-type> #'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("animation-trigger-type", "once");
test_valid_value("animation-trigger-type", "repeat");
test_valid_value("animation-trigger-type", "alternate");
test_valid_value("animation-trigger-type", "state");
test_valid_value("animation-trigger-type", "once, repeat");
test_valid_value("animation-trigger-type", "once, repeat, alternate");
test_valid_value("animation-trigger-type", "once, repeat, alternate, state");
</script>
</body>
</html>
